Adding initial implementation of "tracks" for mixing audio
A "track" in the JSSynth layer corresponds with a track in the UI
layer. It allows setting a master gain for all notes played on the
track. This means the master track volume can be set in one place,
instead of being manually set on all notes played on the track. In
the future, it will also allow adding effects (delay, reverb, etc.)
to all notes played on a track - effects should be added to the
entire track instead of individual notes to acheive the expected
result.

In this commit, a single track is hard-coded, and all notes are
played on this track. Since the track only consists of a gain node
with amplitude 1.0, it is essentially a pass-through. However, this
allows hooking up the basic infrastructure. In future commits, a
full system of multiple tracks will "actual" volumes be implemented.