Track volumes mixed by `AudioSource` tracks, not note volumes
Each note now has its volume set to 1.0, and the actual volume
of the notes played on a track is determined by the track's
separate volume control. This is more elegant, and allows separating
the "sheet music" from the final volume mix. This still allows giving
each note a separate volume, such as in response to a MIDI velocity
attribute, while not taking advantage of that currently.

This commit includes a regression to no longer normalize each track
volume to avoid clipping. That should be re-introduced in a future
commit.

This commit also does not update offline audio generation to use
the new track system, so until that is implemented offline audio
generation will also be broken.