Notes can be scheduled on `AudioSource`
In addition to the existing ability to play them immediately. This
allows `AudioSource` to be a central place to play all notes, instead
of playing scheduled notes in one place (an `Instrument`), but immediate
notes in a different place.

`Instrument` still has the ability to schedule notes for later playback,
to avoid coupling it to `AudioSource`.