Mixer channel amplitudes auto balanced to prevent clipping
This re-introduces the auto-balancing which was previously implemented
before the introduction of mixer channels. Previously, this was handled
by setting the amplitude of individual notes accordingly. Now, it is
handled by scaling the amplitude of each mixer track.

This way of doing things seems better because it decouples the "sheet
music" from the mixer, and also provides a centralized place to balance
the amplitudes, instead of having to be done anywhere notes are triggered.
It also allows balancing to happen automatically inside the mixer,
removing the need for the caller to manually deal with it.