Fixing "download to *.wav" functionality to work again
This isn't the best way to implement this, since it means that
`OfflineTransport` is now coupled to a specific data structure
from the UI layer. This could be resolved by manually converting
the UI track data structure into a more generic format just for
the purpose of creating an `OfflineTransport`, but I'm being lazy
and avoiding that because it seems like something that would be
a short-term solution anyway.

I think the _real_ solution is probably that there is a better
design for `OfflineTransport` that would allow it to take a
previously constructed `AudioSource` instead of constructing it
itself. However, this commit allows the functionality to be
restored in the meantime until a hypothetically better design
is implemented.