Partial support for delay effect on each track
The delay can only be configured on the default song, and can't
be modified after the app is loaded or set for newly created
tracks. But if configured, there will be a delay effect on the
relevant track.

The delay is simple and consists of a delay time plus a feedback
amount, similar to the description on this page:
http://blog.chrislowis.co.uk/2014/07/23/dub-delay-web-audio-api.html

Delay is also ignored when downloading to *.wav, so that should
be fixed in a future commit as well.