Track/instrument delay can be configured via the UI
The method of syncing the delay configuration to the relevant
mixer channel is heavy handed, because it requires re-syncing the
delay for _all_ channels. This is because there's not a good way
to figure out the relevant track from a given instrument. This
might imply that the instrument is not the correct place for
delay configuration to live, or that delay WebAudio nodes should
live on the instrument rather than the track. But I'm not sure
about that. Anyway, the delay is now configurable via the UI,
but there might be a better way to implement it in the future.