Track muting/un-muting now takes effect immediately
The previous implementation of muting worked by removing the track's
notes from the serialized note set, which meant that any notes already
in progress would not be affected by muting/un-muting. This was
noticeable if a long note was being played.

The new muting behavior takes immediate effect, and it also makes more
sense from a code perspective - muting is conceptually part of mixing
the final audio, rather than the "sheet music".