Fixing track volume to not change depending on delay amount
When the delay time was 0, the delay instantly took effect, and
effectively doubled the underlaying base sound. This caused the
following apparent effects:

1.) Tracks with a delay time of 0 were artificially loud
2.) Tracks with a non-zero delay time were quieter than tracks
    with no delay.
3.) Changing the delay time of a track during playback would
    cause the apparent volume of the track to change.

Now, the delay node is only routed (from a logical perspective)
if the delay amount is greater than 0. The end result is that
the presence/absence of delay should not affect a track's volume.