Track muting flag respected for downloaded *.wav output
Previously, all tracks would be played in the downloaded *.wav
file, regardless of whether or not they were muted.

This also future-proofs newly loaded songs to also respect the
"is muted" flag. In the current app this doesn't cause any
behavior to change, because the default song doesn't have any
muted tracks and that's the only song you can load. However, if
a newly loaded song _does_ have muted tracks in the future, it
will now take the "is muted" flag into account.