Avoiding no-op state changes/renders for piano keyboard
Previously, any mouse/touch movement on the piano keyboard would
result in a state change and re-render, even if the set of active
notes didn't change. (E.g. moving the mouse around on the same
keyboard key without moving to a different key). A state change
now only occurs if the active note set actually changes. This seems
to make a noticeable difference in piano keyboard responsiveness,
especially when there are a larger number of measures in the
sequencer, such as 99.