Decoupling score (i.e. notes to be played) from instruments
Notes in the score now only record an abstract channel for them to
be played on, rather than a full `Instrument` object that can be
used to play them back.

In order to implement this, a new `NotePlayer` object is being
introduced. It provides "channels" that a particular note can be
played on in order to be played by a particular instrument. Each
`NotePlayer` channel is then connected to an `AudioSource` channel
in order to be placed into the final mix and have effects applied.
This also allows `AudioSource` to have a more focused purpose of
being a mixer and effects processor, rather than also being a note
player.

In general it seems like the "sheet music" (i.e. the score) should be
de-coupled from the actual performance of the score. This matches how
MIDI works conceptually. It also means that instruments no longer need
to be re-synced when the score changes, and the score no longer needs
to be re-synced when instruments change.