Extracting `Score` object to replace raw array of notes
In the current version it is little more than a wrapper, but it
seems desireable to have the score represented by a first-class
object rather than an array with a specific structure. In the
future, the `Score` object can be fleshed out more so that it is
more than just a pass-through.