Octaves start at C, not A, to match how MIDI works
The original "octave starts on A" behavior was originally done
because it seemed logical to me at the time. However, it would
be better to use MIDI as a reference rather than trying to use
a custom scheme.

I'm not 100% sure if the octave mappings are off by 1 or not.
I.e., should MIDI note 12 be C0, as it currently is, or C1? I've
seen conflicting info when researching this.