Preventing error when very low MIDI note is played
The error is that MIDI notes 0-11 are for the -1 octave, and note
string parsing doesn't work properly for a note such as "A-1". It
might be better to support very low notes, but in the meantime this
at least prevents an error.

One way to fix this would be to shift all of the octave names up
by 1, so that MIDI note 12 refers to C1, not C0.