Downloaded *.wav fades out delay/reverb effects at end of song
The length of the fade out time is constant (1.5 seconds), regardless
of which effects (if any) are activated. This means that a very long
delay might be truncated, or if there are no effects, extra silence
might be added to the end of the song. However, any truncated sounds
sounds will smoothly fade out instead of suddenly stopping.

If it is desired to have long delays be preserved in the output, a
workaround would be to add extra blank measures to the end of the
song to account for this.

It might be possible to have smarter logic for determining the fadeout
time. For example, if no effects are enabled, remove the fade out
entirely. However, for some effects settings such as feedback delay
I'm not sure how to calculate a more exact fadeout time. The current
behavior seems to match how Garageband works, so it doesn't seem like
the worst thing in the world?