Pattern options in sequencer updated when track name changes
The logic to determine if the set of available patterns had
changed was incorrect, causing the drop-downs to not re-render
when a track name changed. This was behavior that previously
worked correctly, but regressed when changing the `TrackMeasure`
component to implement `shouldComponentUpdate()` as part of
af961765c68cf8cd89391d8f42400beb02fbb683.

To fix this, a dedicated data structure is used to define the
available patterns for each track. This ensures that the prop is
passed "by value", so to speak, not reference, so that if the
name changes the "before" and "after" values will be different.
However, this also means the data structure is regenerated on each
app render, so in the future it might be better to cache it somehow.