Playhead head not reset back to beginning when playback stops
This change is in preparation for allowing the playback head to be
dragged to change the current step. In the current UI this behavior
is not ideal because it means there's no easy way to get back to the
start of the song, but that should be resolved by allowing the playback
head to be dragged to a new position. (Or adding a "reset back to
beginning" button in the UI).