Playback head can be dragged to a different step
This is allowed whether or not playback is active. The playback head
now has the appearance of a slider thumb for easier selection, and a
new thin bar extending the height of the sequencer is also lined up
with the current step for an additional visual indicator.

There are some minor visual problems which should be fixed in the future.