Pattern start step stored on pattern itself, not track
An implication of this change is that a given pattern can only exist
once in a song, instead of multiple times. This seems like a worthwhile
tradeoff in exchange for allowing each pattern to start at an arbitrary
position.

To reduce the amount of boilerplate changes required for the demo song,
the various places where patterns repeat are not being preserved. This
means the demo song will be missing a lot of expected content, but it
can be added back in once this general functionality is complete, and
make things easier to work with in the meantime.