Pattern editor doesn't blow up if no pattern is currently selected
This is an intermediate step to further move the editing UI from being
centered around the track editor and instead more around the sequencer.
This commit makes the track editor UI confusing when no pattern is
selected, because the "Patterns" tab will be an empty void. However,
this makes the UI more flexible by not _requiring_ a pattern to be
selected, which will aid in implementing the new sequencer UI, and
also makes some of the code simpler.