Focusing track name field selects the track in track editor
This is an intermediate step toward making the track editor only
activated by performing actions in the sequencer, and removing the
"Instrument" and "Pattern" tabs. Focusing on an instrument by focusing
on the track name field is not the intended final UI, but it allows
bootstrapping things to implement the necessary behind-the-scenes
logic.