Removing track editor; instruments and patterns now edited individually
This removes the notion that a track is selected at all times. Now instead
a track can be selected (in order to edit the attached instrument), a
pattern, or nothing. The pane at the bottom of the page that houses the
instrument or pattern editor will now be empty if neither a track nor a
pattern is selected.

This moves closer to a UI in which the instrument and pattern editors are
shown in separate views linked from the sequencer, instead of an always
visible pane at the bottom of the window.