Instrument editor shown in full-window view
This is the first step toward changing the UI so that the three main
editing panes (sequencer, instrument editor, pattern editor) are shown
in their own full-window views, instead of a top pane for the sequencer
and a bottom pane for the instrument/pattern editors. This seems like
it will work better on smaller mobile screens, and make it easier to
add future potential features.

As part of this, when a new track is added, it is no longer automatically
selected, because otherwise that would cause the instrument editor to be
immediately opened, which seems like confusing behavior.