New UI to select pattern for editing
When a pattern box in the sequencer is clicked, an "Edit" button
floating above the pattern is now shown. Clicking this button will
activate the pattern in the pattern editor.

In the future this is intended to appear as a pop-up bubble, but
this commit provides the basic implementation to make it functional.