Adding button to each track header to edit the instrument
This is a better way of enabling the instrument editor than focusing
on the track name field. A side effect of this is that it's no longer
possible to edit the name of a track in the sequencer. In the future,
this ability should be added to the instrument editor.