Multi-measure pattern doesn't cause app crash when song shortened
When the number of measures is reduced, patterns that are past the new
end of the song should be removed. Previously this removal logic assumed
that all patterns were 16 steps long, which could cause patterns longer
than 16 steps to not be removed. This would then cause an app crash
further downstream.