More specific folder for webpack output
The output folder is now "dist/", rather than the root folder. "dist/"
has always been the intended output folder, and this removes the need
to prefix various paths with "dist/". This should also be safer to use
with `clean-webpack-plugin`, to make it less likely that the root folder
is accidentally deleted.