Patterns with arbitrary length clamped to end of song
The previous logic to clamp patterns to the end of the song assumed
that all patterns were 16 steps long. If a pattern was longer, the
pattern would be allowed to extend past the end of the song, which
would cause an error.

As part of this change, the clamping logic is moved up out of the
component level and into the service layer, which seems like a more
appropriate place for this logic to live.