Fixing build process to always include unchanged files
The build process was accidentally broken when upgrading to
`copy-webpack-plugin` v5.0.3. Starting in v5, this plugin only
copies modified files to the output folder when the app is built
using the `--watch` option. Since the output folder is erased
before each build, this prevented the main *.html file and the
image and sound assets from being included in most builds.