Patterns can be selected with touch inputs
Patterns can't be moved using touch inputs yet, but they can be selected
by touch tapping on them. They can also be edited/removed by tapping again
to make the popup menu visible.

This is a little wonky because the use of a hidden <input> to handle focus
means that automatic scroll-to-focused-element on iOS Safari can cause the
viewport to unexpectedly scroll to the focused pattern. To try to mitigate
this somewhat, when selecting a pattern the hidden input is moved to be
adjacent to it, which reduces the amount of scrolling. However, this doesn't
seem like a perfect solution, so a different approach to handling pattern
focus might be needed.