Clamping patterns to start of a measure
This limitation is intended to make it easier to implement certain
features such as adding a new pattern without overlapping an existing
pattern. Although it would be nicer to allow patterns to be dragged
to arbitrary steps or to have smart merging when patterns overlap,
these limitations make it easier to launch a "draggable patterns" MVP
in the meantime.