Fixing optimization to only re-render when pattern is moved
When a pattern is dragged with the mouse, many drag events can be
quickly created. To avoid unnecessary React renders, the state should
only be updated when the pattern is actually dragged to a different
track or start step.

The previous implementation of this wasn't correct, because it relied
on a non-existent prop (I believe due to code that moved to a different
component without being updated).

This new implementation should actually be correct, and is also
refactored to move the optimization to the core `movePattern()` service
function, instead of the view component. This seems like a better place
for this to live, since it will allow the optimization to occur any time
a pattern is moved, regardless of the UI used to do so.