More robust method of clamping pattern to bounds while dragging
The clamping now occurs in the UI layer, by clamping the raw pixel
coordinate of the mouse/touch input, rather than at the service
layer by clamping the step index/track index.

This seems a little better because it seems a little strange for
the service layer to clamp to step index/track index rather than
throwing an error.

This is a step toward implementing the ability to drag patterns using
touch - the relevant clamping isn't fully necessary when using the
mouse, because mouse events will be clamped to the container anyway,
unlike touch events.