Fixing variable to stay in sync with number of measures
The logic introduced in 2c854994ab63a05e756eeb75254e3412d79d9e0a to
add `this.timelineWidthInPixels` was flawed because this variable
was only set when the component was initialized, and didn't stay in
sync with the relevant props. In practice, this meant that if the
number of measures was changed, the timeline display would become
messed up due to this variable having an obsolete value.