Fixing pattern dragging across tracks using mouse in Firefox
Firefox v68 seems to give a different value for the `buttons` attribute
of the `mouseover` event compared to Safari and Chrome when the event
occurs on a child element of the container (specifically it gives a
`buttons` value of 0, instead of 1). This causes the drag action to
be prematurely ended when dragging a pattern to a new track.

The `mouseenter` event seems to be handled the same by Firefox, Safari,
and Chrome, and it also has the virtue of being only called when the
parent container is entered, instead of its children. This results in
fewer events to handle which maybe is slightly better for performance
(although that is a secondary concern).