Adding specific field to indicate pattern step count
This allows the size of the pattern shown in the UI/used during playback
to differ from the pattern data in the underlaying data structure. This
is intentional, because it allows e.g. a pattern to be truncated in the
UI without throwing away the hidden pattern data if the user desires to
re-expand the pattern in the future.

This also fixes the process of adding a new pattern row to include the
correct number of steps, instead of always using 16 steps for the new
row. This was a bug introduced in an earlier change when adding support
for patterns that aren't hard-coded to 16 steps long.

This is being done as a preliminary step toward allowing patterns to be
resized in the UI.