Allowing the length of patterns to be changed in the UI
The length of the pattern can be changed by dragging the right edge
of the pattern. This is similar to how patterns can be resized in
GarageBand/Logic.

Currently the length of each pattern is clamped to be a multiple of
a full measure. This restriction only exists in the UI layer, and not
the underlaying pattern data structure.