Placeholder rendering of pattern with looped sections
This visually approximates how these patterns should look, as a
placeholder to help implement the ability to loop sections in the
UI. The primary shortcoming is that each "sub pattern" (for lack
of a better term) should only have a soft visual boundary, instead
of a hard visual boundary. That can be fixed in a future commit.