Basic ability to loop a pattern in the UI
A pattern can be looped by dragging an area on the right side of the
pattern box, similar to how it can be lengthened/shortened. This UI is
very similar to how it works in GarageBand/Logic.

Currently patterns can only be looped in increments of a full measure
to simplify the implementation.