Adding basic ability to edit track name in instrument editor
This is a placeholder UI, which shows an input textbox that allows the
name to be edited. In the future, the intention is to replace this with
a read-only field that can be toggled to show a textbox for editing along
with a "Save" button. (Similar to the current UI to change the number of
measures in the song).

It seems like maybe the name should live on the instrument rather than
the track, but not dealing with that right now.