Adding basic ability to copy and paste patterns in the sequencer
The specific UI is a placeholder - a new "Copy" button is added to the
selected pattern popup menu, and a copied pattern can be pasted to a
given measure by clicking while holding the ALT key. The main drawback
of this UI is that it doesn't support mobile devices (similar to how
the current UI to add a new pattern doesn't either).

Pasted patterns aren't checked for overlap with other existing patterns,
so patterns can be pasted on top of each other. This will be prevented
in a separate commit.