Better variable names that more directly describe what values represent
These two state values are really used to set the position of the relevant
popup menu, and the fact that this position is calculated using the location
of a selected pattern box is incidental to the final value.

This is a being done in advance of adding a new type of popup menu.