More flexible way of calculating popup menu vertical position
This allows a sequencer popup menu to be placed without reference to
an existing pattern box. This is a step toward allowing a popup menu
to be shown when clicking empty space in the sequencer (in order to
add/paste a pattern), since there won't be an existing pattern box
on which to base the position.