Decoupling setting selected pattern from setting popup menu position
This makes it easier to modify the popup menu without having to also
deal with highlighting a certain pattern, and is a step toward the
ability to show a popup menu outside of the context of a selected
pattern. This also allows removing the brief relocation of the popup
menu to the top-left corner of the sequencer when starting a resize
drag of a pattern.