Adding popup menu to sequencer timeline to add/paste patterns
This replaces the ability to do this by clicking in the timeline with
a modifier key pressed. Besides this new interface seeming like it is
more discoverable, it also works with touch devices.

This implementation works on mobile, but via defaulting to the
`mousedown` event, instead of with actual touch events. This is being
done because it's easier to implement it this way, but in the future
it might be better to update the implementation to use actual touch
events.