Patterns are copied by value, not by reference
Previously, when copying a pattern, only the pattern ID was stored,
and when pasting, the current version of the copied pattern was duplicated,
not the version of the pattern in place when the copy request happened.

This change seems like more expected behavior and is how GarageBand/Logic
work. It also allows copied patterns to be pasted after the original pattern
has been removed.