Smarter placement of sequencer popup menu when near window edges
This prevents the popup menu from being shrunken when triggered near
the right window edge (even though there is enough room for the menu),
and also allows the arrow below the bubble to always be placed above
the target point, even if it makes the arrow off center.

There still seems to be an issue on iOS with the "Add/Paste" menu being
shrunken near the right edge, which should be fixed in a future commit.
In addition, when the popup menu is triggered very close to the window's
right edge, the bubble arrow can be disconnected from the bubble. This
should be fixed in the future as well.