Sequencer popup menu now hidden when scrolling the timeline
This avoids either the popup menu being in the wrong location (due to
it not moving with the rest of the scrolling), or having to reposition
the popup menu with each scroll event. In practice, since scroll events
can happen rapidly, the latter would require many state changes for
seemingly little benefit.