Preventing popup menu when clicking timeline outside song bounds
Previously, clicking on the timeline in the empty space before/after
the main song content would show the "Add/Paste" popup menu, but with
the position clamped to the beginning/end of the song. Since it's not
possible to add patterns outside of the normal song bounds, and since
the popup menu will probably not be placed where the user intended,
showing this popup menu in this situation doesn't seem to make sense.