Truncating song doesn't cause crash if pattern loops past new end
Previously, patterns whose base length extended past the new end of
the song did not cause a problem, but a pattern whose looped portion
extended past the new end would cause a crash.

It seems a little wonky that when truncating a song patterns that
extend past the new end will be fully removed instead of being
truncated, but this at least prevents the app from crashing. Better
behavior can be implemented in a future commit.