New pattern names no longer given numeric suffix
This is incremental progress toward having pattern names be set by
the user, instead of automatically generated.

The numeric suffixes were an artifact of the old sequencer/pattern
editor UI in which patterns could not be given custom names. The
numeric suffixes provided the only way to tell patterns apart. In
the current UI, these suffixes seem unnecessary and arbitrary.