Preventing "Add/Paste" menu when clicking in pattern box "dimples"
Before this commit, clicking in the whitespace caused by the curved
border radius between each looped section of a pattern would trigger
the "Add/Paste" menu, instead of highlighting the relevant pattern.
This would allow adding a new pattern on top of an existing pattern,
which is not supposed to be allowed. In practice this was a little
hard to do since the whitespace is not very large, but it was still
possible.

After this commit, clicking the whitespace will instead highlight
the relevant pattern.