Better visual appearance for looped pattern sections
Each sub-box in a looped pattern is now flush with the other sub-boxes
in the parent pattern, and separated by a dashed line. (Before, they were
separated by 1px and had solid left/right borders). This seems to make it
more clear that the sub-boxes are part of the same pattern, rather than
independent patterns.