Long pattern names can expand into looped pattern boxes
Previously, when a pattern contained looped sections, the pattern
name would only be shown in the first pattern box, and be truncated
if it was too long. Now the pattern name can take up the entire
width of the pattern.

(However, if the name takes up the full width of the pattern it will
overlap with the "loop" button, so that should be fixed in a future
commit).