More intuitive pattern dragging behavior
Previously, each drag event would cause the pattern to be moved such
that it started at the beginning of the measure under the pointer.
However, this could cause large jumps when e.g. clicking near the
right edge of a long pattern and dragging straight down. When moved
to a new track, the pattern would suddenly start multiple measures to
the right of where it originally started, which seems unexpected.

Now, dragging is relative to the measure in which the drag started.
In the scenario above, i.e. clicking near the right edge of a pattern
and dragging perfectly straight down, the pattern's start step will
not change when dragging the pattern to a new track. After trying this
out it seems easier to use, and is similar to how dragging works in
GarageBand/Logic.