Preventing unnecessary renders when dragging patterns in sequencer
Previously, while a drag/resize/loop change was in progress the
sequencer timeline was re-rendered upon each pointer move event.
Since most of these move events did not cause any of the timeline
content to change, this was unnecessary.

Now the timeline should only be re-rendered when a pointer move event
actually causes the timeline content to change.