Pattern resize button always visible, but sometimes disabled
Previously, if a pattern could not be resized because it had
been looped, the "resize" button was hidden. Now, it will always
be shown, but grayed out in this scenario. This seems like more
clear behavior.

However, the grayed out behavior will currently only happen if
the reason is that the pattern has been looped - if the pattern
can't be resized because it's next to another pattern or is at
the end of the song, the button will not have a grayed out
appearance. This matches the existing behavior, and also the
behavior of the "loop" button. This could possibly be changed
in the future, but seems out of scope for this commit.