Converting <TimelinePattern> to pure component
This greatly reduces how often timeline pattern boxes are
re-rendered. This is especially helpful when state changes
happen rapidly but don't directly affect the pattern boxes.
For example, during song playback, while dragging the playback
head, or dragging a slider to change a track's volume.

However, enabling/disabling the timeline popup menu will still
cause every pattern box to be re-rendered, due to how that is
implemented. That can be fixed in a future commit.