Using single field to indicate current drag mode
The timeline container can only be in a single drag mode at a
time, but previously 3 different variables were used to indicate
the current mode. The idea was that only one of the variables
should be set to `true` at a time, and that variable would
indicate the current drag mode.

However, if more than one _was_ set to `true` (e.g. if there was
a bug) that would cause problems. Using a single field eliminates
that possibility, is easier to understand, and is more succinct.