Removing unnecessary intermediate data conversion
Before this commit, `Serializer.serializePatternRows()` took an
array of note strings, converted it to a joined string, passed it
to `SynthCore.SequenceParser.parse()`, which finally split the
string to get back the original array of note strings! To remove
this unneeded intermediate step, `SynthCore.SequenceParser.parse()`
now takes an array of note strings directly.