Preventing extra no-op notes in parsed note sequence
Because empty notes are represented by "" and not " ", the
expected branch to handle empty notes was never used. This
resulted in extra notes with invalid note name/octave values
being present in the parsed sequence.

Since these notes were excluded from the serialized score
further down the line, this didn't cause any direct problems
in practice. However, it's better to avoid this unnecessary
data noise, and preventing this type of invalid note will
also make it easier to implement proper filtering of _actual_
notes with invalid values (e.g. "@3") in the future.