Invalid pattern notes properly ignored during playback
Previously, an invalid note entered in the pattern editor such as
"@1" would cause a console log message to appear in Firefox and
Chrome, and in Safari a note with 440Hz frequency to be included
in the song.

The reason that Safari added a 440Hz note was that unlike Firefox
and Chrome, Safari (as of v13.0.5) doesn't throw an error if an
`AudioParam.value` is set to an invalid value such as `NaN`.
Instead, the invalid value is silently ignored. Since an
`OscillatorNode` has a default frequency of 440Hz, and attempting
to set the frequency to an invalid value didn't result in an error
being thrown, a note with the default 440Hz frequency was added.

To implement this change, the `SynthCore.Note()` builder function
now intentionally blows up if an invalid value is passed in for
one of the arguments, and it is more strict about what counts as
valid. It seems like a better design to blow up in the case of bad
input rather than have an object with invalid values that later
needs to be dealt with. For example, this allows removing the
check in `Serializer.serializeScore()` that each note is valid
before adding it to the score.