Fixing *.wav download on iPhone
The general fix is to make the `OfflineAudioContext` used to
create the *.wav file have the same sample rate as the main
`AudioContext`, instead of being hard-coded to 44,100Hz.

To explain the problem, first two pieces of background info:

1.) If a `ConvolverNode` created with one sample rate has an
    `AudioBuffer` with a different sample rate assigned to its
    `buffer` property, an error will occur.

2.) Safari/Firefox/Chrome on macOS and Safari on iPadOS use a
    44,100Hz sample rate for the main `AudioContext`, but Safari
    on iOS (as of iOS 13) uses a sample rate of 48,000Hz.

`ConvolverNode` instances used for creating the *.wav file are
created from the `OfflineAudioContext`, and therefore had a
sample rate of 44,100Hz. However, `AudioBuffer` instances
assigned to the `buffer` property are created from the main
`AudioContext`, and on iOS have a sample rate of 48,000Hz. This
difference was the source of the error.

Since browsers on macOS and iPadOS use a 44,100Hz sample rate
for the main `AudioContext`, everything had the same sample rate
and no errors occurred.