Popup menu hidden when sequencer "is expanded" changes
When the "is expanded" state changes, the sequencer rows shift
position, but an active popup menu does not move along with them.
This results in the menu being in the wrong place visually.

The simplest fix is to hide the menu in this scenario, since it
can always be re-shown by the user if they want. This is analogous
to how the menu is also hidden when the sequencer is scrolled.