Keyboard notes stopped when instrument/pattern editor closes
Previously, when the instrument or pattern editor was active, and
a note was triggered by an external MIDI keyboard, and held while
the user navigated back to the main sequencer view, the note would
continue to play indefinitely even after the note was stopped on
the MIDI keyboard. It would not stop until the user navigated back
to the instrument or pattern editor and manually re-stopped the
note (either by re-tapping the note on the on-screen keyboard, or
by re-playing and stopping the note on the MIDI keyboard).

In theory this could have also happened for notes triggered via
the on-screen keyboard, but in practice clicking the "back" button
requires taking the mouse/touch points off the keyboard (and hence
stopping any notes being played).