Simplifying markup of sequencer timeline grid
This implements the timeline grid as a single container element,
with pattern boxes inside positioned absolutely both horizontally
and vertically. Previously, the container element contained a list
with an item for each track, and each pattern box was positioned
inside one of the list items.

This is being done in advance of implementing a new way of
focusing/un-focusing pattern boxes. For the focus to be preserved
correctly when dragging a pattern to a new row the pattern boxes
all need to be inside the same container. I think this has
something to do with how React re-creates/preserves elements on
the page, and maybe there is some React way to work around that,
but the new markup seems like a better implementation regardless.