Better way of focusing/un-focusing sequencer pattern boxes
Normally only certain types of elements are focusable, but if an
element (regardless of type) has the `tabIndex` attribute set to
an integer it becomes focusable. This is a simpler way of focusing
and un-focusing sequencer pattern boxes than the previous way of
indirectly using a hidden `<input>` element to simulate this.

The `tabIndex` for pattern boxes is intentionally set to a
negative value because a value 0 or greater allows using the TAB
key to tab between elements. That might be a nice thing to support
in the future, but for now preserving the existing behavior.