Converting `<NoteBox>` to be a pure component
This avoids a lot of unnecessary renders. Previously, when an
individual note box was changed, _all_ other note boxes would be
re-rendered as well.

Now that `<NoteBox>` is a pure component, for changes to be
re-rendered the entire underlaying note object has to be replaced,
instead of only the note name as was done before.