Preventing adding pattern on top of existing pattern
Previously, clicking on the 1px gap between pattern boxes would
show the "Add/Paste" popup menu, and clicking "Add" would allow
adding a new pattern on top of an existing pattern.

After this commit, the pattern boxes look the same (they visually
still have a 1px gap between them), but the container elements
which are their click targets are 1px wider and taller. When
clicking in a 1px visual gap, this ensures that one of the pattern
boxes is selected, instead of showing the "Add/Paste" menu.